// assets/script/protobuf/holdem/code_pb.d.ts
export interface CodeMap {
  NONE: 0;
  MSG_D_REGISTER: 1;
  MSG_D_HEARTBEAT: 2;
  MSG_S_ERROR: 99;
  MSG_S_NOTIFICATION_ROOM_READY: 101;
  MSG_S_USER_BAN: 102;
  MSG_S_NOTIFICATION_SYSTEM_MAINTAIN: 103;
  MSG_S_ROOM_READY_FOR_ENTER: 104;
  MSG_S_MTT_READY_FOR_APPLY: 105;
  MSG_S_USER_KICKED_FROM_CLUB: 106;
  MSG_S_NOTIFICATION_MTT_WILL_START: 107;
  MSG_S_SYSTEM_MESSAGE: 108;
  MSG_S_GET_MESSAGE: 109;
  MSG_S_USER_GAME_WATCH: 110;
  MSG_S_FRIEND_ROOM_BRING_IN_APPLY_TO_ADMIN: 111;
  MSG_S_FRIEND_ROOM_BRING_IN_APPLY_TO_USER: 112;
  MSG_S_CLUB_ROOM_BRING_IN_APPLY_TO_ADMIN: 113;
  MSG_S_CLUB_ROOM_BRING_IN_APPLY_TO_USER: 114;
  MSG_S_CLUB_ROOM_BRING_IN_APPLY_AUDIT: 116;
  MSG_S_ADMIN_ROOM_USER_STANDUP: 117;
  MSG_S_ADMIN_ROOM_USER_LEAVE: 118;
  MSG_S_ROOM_DELAY_APPLY_TO_ADMIN: 119;
  MSG_S_CLUB_ROOM_DELAY_APPLY_AUDIT: 120;
  MSG_S_USER_IS_BLOCKED: 121;
  MSG_S_CLUB_USER_IS_BLOCKED: 122;
  MSG_S_TODO_LIST: 123;
  MSG_S_USER_DEVICE_IS_BLOCKED: 124;
  MSG_S_OFFLINE_TICKETS: 125;
  MSG_S_SNATCH_TREASURE_HRL: 126;
  MSG_S_SNATCH_TREASURE_WIN_POPUP: 127;
  MSG_S_JACKPOT_MARQUEE: 128;
  MSG_S_MATCHING_RESULT: 129;
  MSG_S_SELF_PROFIT_PAY: 130;
  MSG_S_LIMIT_HAND_NUMBER: 131;
  MSG_S_ROOM_USER_SEND_DIAMOND: 132;
  MSG_S_USER_ORDER_AUDIT: 133;
  MSG_S_USER_WHEEL_HAND_NUM: 134;
  MSG_S_CACHE_DATA_UPDATE: 135;
  MSG_S_SUPPORT_MESSAGE: 136;
  MSG_S_USER_IS_MUTE: 137;
  MSG_S_USER_DIAMOND_CHANGE: 138;
  MSG_S_TRIBE_BLACK_USER_MTT: 139;
  MSG_S_ROOM_CHANGE_NOTIFY: 140;
  MSG_S_USER_GOLD_CHANGE: 141;
  MSG_S_TRIBE_BLACK_USER: 142;
  MSG_S_USER_TRADER_ORDER_NOTIFY: 143;
  MSG_S_ROOM_MTT_SETTLE_NOTIFY: 144;
  MSG_S_USER_USDT_ORDER_NOTIFY: 145;
  MSG_S_FRIEND_ROOM_CREATOR_SETTLE: 146;
  MSG_S_MTT_AWARD_NOTIFY: 147;
  MSG_S_USER_JOIN_CLUB: 148;
  MSG_S_CLUB_ROOM_MTT_SETTLE_NOTIFY: 149;
  MSG_S_USER_CLUB_ROLE_CHANGE: 150;
  MSG_S_USER_MTT_CHANGE_NOTIFY: 151;
  MSG_S_USER_SNG_CHANGE_NOTIFY: 152;
  MSG_S_MTT_SERIES_NOTIFY: 153;
  MSG_S_MTT_BREAK: 154;
  MSG_D_USER_PLAYING: 201;
  MSG_R_ROOMS: 202;
  MSG_R_QUICK_JOIN: 203;
  MSG_R_MTT_DETAIL: 204;
  MSG_R_JOIN_MATCHING: 205;
  MSG_R_ROOMS_SIMPLE: 206;
// ...
}

export const Code: CodeMap;